Feat: 게시글 읽기, 삭제, 수정 구현

+++ src/main/java/leets/weeth/domain/post/service/PostService.java
@@ -3,11 +3,12 @@ package leets.weeth.domain.post.service;
 import jakarta.transaction.Transactional;
 import leets.weeth.domain.post.dto.PostDTO;
 import leets.weeth.domain.post.entity.Post;
-import leets.weeth.domain.user.entity.User;
 import leets.weeth.domain.post.repository.PostRepository;
+import leets.weeth.domain.user.entity.User;
 import leets.weeth.domain.user.repository.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -24,9 +25,13 @@ public class PostService {
         return postRepository.findAll();
     }
 
+    public Post show(Long postId) {
+        return postRepository.findById(postId).orElse(null);
+    }
 
-    public Post create(String userEmail, PostDTO dto) {
-        User user = userRepository.findByEmail(userEmail).orElseThrow(()->new
+
+    public Post create(String email, PostDTO dto) {
+        User user = userRepository.findByEmail(email).orElseThrow(()->new
                 IllegalArgumentException("failed to add post! no such user"));
         Post post = Post.createPost(dto, user);
         Post created = postRepository.save(post);
@@ -34,21 +39,28 @@ public class PostService {
     }
 
     @Transactional
-    public PostDTO update(Long postid, PostDTO dto) {
+    public Post update(Long postid, PostDTO dto, String currentEmail) {
+
         Post target = postRepository.findById(postid)
                 .orElseThrow(()->new IllegalArgumentException("failed to edit the Post. so such post."));
+        if (!target.getUser().getEmail().equals(currentEmail)) {
+            throw new AccessDeniedException("You do not have permission to edit this post");
+        }
         target.patch(dto);
         // 2. post 수정
         Post updated = postRepository.save(target);
         // 3. DB로 갱신
-        return PostDTO.createPostDTO(updated);
+        return updated;
 
     }
     @Transactional
-    public PostDTO delete(Long postid) {
+    public Post delete(Long postid, String currentEmail) {
         Post target = postRepository.findById(postid)
                 .orElseThrow(()->new IllegalArgumentException("no such post"));
+        if(!target.getUser().getEmail().equals(currentEmail)){
+            throw new AccessDeniedException("You do not have permission to delete this post");
+        }
         postRepository.delete(target);
-        return  PostDTO.createPostDTO(target);
+        return  target;
     }
 }
